Add a new activity for the roulette game board.
The goal here was to setup the layouts and a couple of simple widgets to achieve the look of a roulette game board. In the component tree view of Android Studio, you can see how different layouts have been nested to achieve the roulette board layout. The widgets themselves can be modified later on when we have established which ones would serve best to achieve the look and feel we need.