Add source for a simplified blackjack game.
A single instance of the BlackJack class can be used to play the game repeatedly. The user of the class supplies a bet amount when calling the play() method. The player can then choose to perform a "hit" action in which they draw a card from the deck. The player may also choose to "stand" in which case the game ends and a payout amount is calculated based on whether the machine or player has won the game and the bet amount. Note, the player's hand and machine's most recently drawn card is always accessible for viewing as per standard blackjack rules.

SAMPLE USAGE:

import java.util.Scanner;

public class Test {
    private static void printGameState(final BlackJack bj) {
	System.out.println("Dealer Card: " + bj.dealerTopCard());
	System.out.println("Player Hand: ");
	for (Card c : bj.playerCards()) {
	    System.out.println(c);
	}
    }

    public static void main(final String[] args) {
	Scanner scanner = new Scanner(System.in);
	BlackJack bj = new BlackJack();
	bj.play(20);

	int payout = 0;
	while (0 == payout) {
	    printGameState(bj);

	    int hit = Integer.parseInt(scanner.next());
	    if (1 == hit) {
		bj.hit();
	    } else {
		payout = bj.stand();
	    }
	}
	System.out.println("EARNINGS: " + payout);
    }
}